requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
mysql-connector-python==9.3.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
gunicorn==20.1.0
Werkzeug==3.1.3
PyMySQL==1.1.1
psycopg2-binary==2.9.10
supabase==2.15.3
